rf: move internal modules to appropriate places
