+++ route4me/sdk/endpoints/optimizations.py
@@ -23,7 +23,6 @@ a number of routes. (Possibly recurring in the future)
 
 import pydash
 
-from .._net import NetworkClient
 from ..models import Optimization
 from ..enums import OptimizationStateEnum
 
@@ -32,6 +31,7 @@ from route4me.sdk.utils import PagedList
 
 from route4me.sdk._internals import add_limit_offset_to_query_string
 from route4me.sdk._internals.typeconv import bool201
+from route4me.sdk._internals.net import NetworkClient
 
 
 class Optimizations(object):
